reuse enums
